+++ recursive_diff/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from recursive_diff.tests import requires_dask
+
+
+@pytest.fixture(params=[False, pytest.param(True, marks=requires_dask)])
+def chunk(request):
+    return request.param